modify features

+++ src/main/java/com/movie/backend/service/MovieService.java
@@ -6,8 +6,10 @@ import com.movie.backend.dto.GenreDTO;
 import com.movie.backend.dto.MovieDTO;
 
 import com.movie.backend.entity.*;
+import com.movie.backend.exception.BadRequestException;
 import com.movie.backend.exception.MovieException;
 import com.movie.backend.exception.UserException;
+import com.movie.backend.repository.EventRepository;
 import com.movie.backend.repository.GenreRepository;
 import com.movie.backend.repository.MovieRepository;
 
@@ -41,6 +43,9 @@ public class MovieService {
 
     @Autowired
     private GenreRepository genreRepository ;
+
+    @Autowired
+    private EventRepository eventRepository;
 
     @Autowired
     private ModelMapper modelMapper;
@@ -208,6 +213,10 @@ public class MovieService {
 
     public void deleteMovie(Long movieId) {
         Movie movie = movieRepository.findById(movieId).orElseThrow(() -> new MovieException("movie not found")) ;
+        List<Event> events = eventRepository.findByMovie(movieId);
+        if (events.size() > 0) {
+            throw new BadRequestException("This movie was belong to event");
+        }
         movieRepository.delete(movie);
     }
     public List<MovieDTO> findBeforeDate() {
